kittens_exhibition/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12
djangorestframework-simplejwt>=5.2.2
drf-yasg>=1.20.0
PyJWT>=2.6.0
pytest
pytest-django
